Duplicates prompt exercises with Gemini
Initial reflections on differences on the model is that Gemini 2.5-flash is more verbose, it leaves explicit instructions in place when creating a list of steps.

It was better at waxing poetic within the few-shot prompting.

Gemini was better at spotting the incorrect student solution, I guess it already forces itself to use more time to reason through the problem first.

Added more observations inline in the Notebook